Delete multiple files

+++ src/main/java/org/embulk/input/aurora_mysql_extract_files/AuroraMySQLExtractFilesFileInputPlugin.java
@@ -1,18 +1,15 @@
 package org.embulk.input.aurora_mysql_extract_files;
 
-import java.io.FileInputStream;
 import java.io.InputStream;
-import java.util.ArrayList;
-import java.util.List;
-
 import java.sql.Connection;
 import java.sql.DriverManager;
 import java.sql.Statement;
+import java.util.List;
+import java.util.stream.Collectors;
 
+import com.amazonaws.services.s3.model.DeleteObjectsRequest;
 import com.google.common.base.Optional;
-import com.google.common.collect.ImmutableList;
 
-import org.apache.http.client.entity.InputStreamFactory;
 import org.embulk.config.Config;
 import org.embulk.config.ConfigDefault;
 import org.embulk.config.ConfigDiff;
@@ -27,18 +24,15 @@ import org.embulk.spi.FileInputPlugin;
 import org.embulk.spi.TransactionalFileInput;
 import org.embulk.spi.util.InputStreamTransactionalFileInput;
 
-import com.amazonaws.auth.AWSCredentials;
 import com.amazonaws.auth.AWSStaticCredentialsProvider;
 import com.amazonaws.auth.BasicAWSCredentials;
 import com.amazonaws.regions.Regions;
 import com.amazonaws.services.s3.AmazonS3;
-import com.amazonaws.services.s3.AmazonS3Client;
 import com.amazonaws.services.s3.AmazonS3ClientBuilder;
 import com.amazonaws.services.s3.model.DeleteObjectRequest;
 import com.amazonaws.services.s3.model.GetObjectRequest;
 import com.amazonaws.services.s3.model.ListObjectsRequest;
 import com.amazonaws.services.s3.model.ObjectListing;
-import com.amazonaws.services.s3.model.Region;
 import com.amazonaws.services.s3.model.S3Object;
 import com.amazonaws.services.s3.model.S3ObjectSummary;
 
@@ -47,6 +41,7 @@ import org.slf4j.Logger;
 public class AuroraMySQLExtractFilesFileInputPlugin implements FileInputPlugin {
 
     private static final Logger log = Exec.getLogger(AuroraMySQLExtractFilesFileInputPlugin.class);
+    private AmazonS3 client;
 
     public interface PluginTask extends Task {
         @Config("aws_access_key_id")
@@ -99,9 +94,23 @@ public class AuroraMySQLExtractFilesFileInputPlugin implements FileInputPlugin {
 
         // delete s3
         try {
+            client = newS3Client(task);
             log.info("deleting objects");
-            AmazonS3 client = newS3Client(task);
-            client.deleteObject(new DeleteObjectRequest(task.getS3Bucket(), task.getS3PathPrefix()));
+            List<String> s3Keys = getS3Keys(task);
+            if (s3Keys.isEmpty()){
+                log.info("no objects are detected for delete");
+            }else{
+                List<String> s3Paths = s3Keys.stream().map(k -> String.format("s3://%s/%s", task.getS3Bucket(),k)).collect(Collectors.toList());
+                log.info("following files will be deleted\n{}", String.join("\n",s3Paths));
+                client.deleteObject(new DeleteObjectRequest(task.getS3Bucket(), task.getS3PathPrefix()));
+                DeleteObjectsRequest multiObjectDeleteRequest = new DeleteObjectsRequest(task.getS3Bucket());
+                List<DeleteObjectsRequest.KeyVersion> keys = s3Keys.stream()
+                        .map(DeleteObjectsRequest.KeyVersion::new)
+                        .collect(Collectors.toList());
+                multiObjectDeleteRequest.setKeys(keys);
+                client.deleteObjects(multiObjectDeleteRequest);
+            }
+
         } catch(Exception e){
             log.error("delete error: ", e);
             return null;
@@ -146,18 +155,11 @@ public class AuroraMySQLExtractFilesFileInputPlugin implements FileInputPlugin {
             return;
         }
 
-        AmazonS3 client = newS3Client(task);
 
         // TODO: read manifest
-        ImmutableList.Builder<String> builder = ImmutableList.builder();
-        ListObjectsRequest request = new ListObjectsRequest().withBucketName(task.getS3Bucket())
-                .withPrefix(String.format("%s.part",task.getS3PathPrefix()));
-        ObjectListing list = client.listObjects(request);
-        List<S3ObjectSummary> objects = list.getObjectSummaries();
-        log.info(String.format("%d files", objects.size()));
-        objects.forEach(object -> builder.add(object.getKey()));
-        List<String> files = builder.build().asList();
-        task.setFiles(files);
+        List<String> s3Keys = getS3Keys(task);
+        log.info(String.format("%d files", s3Keys.size()));
+        task.setFiles(s3Keys);
     }
 
     @Override
@@ -214,5 +216,13 @@ public class AuroraMySQLExtractFilesFileInputPlugin implements FileInputPlugin {
             .standard()
             .withCredentials(new AWSStaticCredentialsProvider(awsCreds))
             .withRegion(Regions.AP_NORTHEAST_1).build();
+    }
+
+    private List<String> getS3Keys(PluginTask task){
+        ListObjectsRequest request = new ListObjectsRequest().withBucketName(task.getS3Bucket())
+                .withPrefix(String.format("%s.part",task.getS3PathPrefix()));
+        ObjectListing list = client.listObjects(request);
+        List<S3ObjectSummary> objects = list.getObjectSummaries();
+        return objects.stream().map(S3ObjectSummary::getKey).collect(Collectors.toList());
     }
 }
